requirements_versions.txt
torchsde==0.2.5
safetensors==0.3.1
accelerate==0.21.0
pyyaml==6.0
Pillow==9.2.0
scipy==1.9.3
tqdm==4.65.0
psutil==5.9.5
pytorch_lightning==1.9.4
omegaconf==2.2.3
gradio==3.41.2
pygit2==1.12.2
opencv-contrib-python==4.8.0.74
diffusers==0.25.1
httpx==0.24.1
onnxruntime==1.17.0
timm==0.9.2
translators==5.8.9
rembg==2.0.53
groundingdino-py==0.4.0
einops==0.7.0
transformers
torch
torchvision
numpy
scipy
matplotlib
scikit-image
basicsr
pillow
opencv-python
insightface
torch
diffusers
transformers
safetensors
pillow
numpy
gdown
realesrgan
basicsr
